requirements.txt
pydantic~=2.11
ipython~=9.6
gradio~=5.49
python-dotenv~=1.2
nest-asyncio~=1.6
playwright~=1.55
openai~=2.7
pytest==8.4.2
langgraph
langgraph-prebuilt
langchain-core
langgraph-cli
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-openai
notebook
langchain-tavily
langgraph-cli[inmem]